test: rename clean.ts to cache.ts
