+++ test/index.ts
@@ -4,4 +4,4 @@ import './install'
 import './uninstall'
 import './link'
 import './prune'
-import './clean'
+import './cache'
